Reorganized component directory for better organization
