+++ frontend/src/components/display/CatalogueList.mjs
