+++ frontend/src/components/display/CatalogueRow.mjs
