+++ frontend/src/components/display/CompositionList.mjs
