+++ frontend/src/components/display/InstrumentationDisplay.mjs
