+++ frontend/src/components/forms/QueuedFeaturedInstrument.mjs
